committed by amata1219.

+++ src/amata1219/hogochi/byebye/RegionByebye.java
@@ -259,7 +259,7 @@ public class RegionByebye {
 		return rb.sales.get(region.getId());
 	}
 
-	public static long getNeedTickets(ProtectedRegion region){
+	public static int getNeedTickets(ProtectedRegion region){
 		if(!isAdminRegion(region))
 			return -1;
